Move sKCSD utility functions to sKCSD_utils
