+++ figures/kCSD_properties/skcsd_and_l_curve_complex_morphology.py
@@ -6,7 +6,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 import sys
 from kcsd import sKCSD, sKCSDcell
-import kcsd.utility_functions as utils
+import kcsd.sKCSD_utils as utils
 import kcsd.validation.plotting_functions as pl
 sys.path.insert(1, os.path.join(sys.path[0], '../sKCSD_paper'))
 import sKCSD_utils
@@ -213,15 +213,15 @@ if __name__ == '__main__':
     morphology, extent = cell_itself.draw_cell2D()
     extent = [ex*1e6 for ex in extent]
     draw_somav(ax_somav, time, somav)
-    k = sKCSD(data.ele_pos,
-              data.LFP,
-              data.morphology,
-              n_src_init=n_src,
-              src_type='gauss',
-              lambd=lambd,
-              exact=True,
-              R_init=R,
-              sigma=0.3)
+    # k = sKCSD(data.ele_pos,
+    #           data.LFP,
+    #           data.morphology,
+    #           n_src_init=n_src,
+    #           src_type='gauss',
+    #           lambd=lambd,
+    #           exact=True,
+    #           R_init=R,
+    #           sigma=0.3)
     path = os.path.join(data_dir, 'lambda_%f_R_%f_n_src_%d' % (lambd, R, n_src))
     if sys.version_info < (3, 0):
         path = os.path.join(path, "preprocessed_data/Python_2")
