+++ figures/sKCSD_paper/make_fig_5.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 import sys
 import os
 from kcsd import sKCSD, KCSD3D, sKCSDcell
-import kcsd.utility_functions as utils
+import kcsd.sKCSD_utils as utils
 import kcsd.validation.plotting_functions as pl
 import sKCSD_utils
