+++ kcsd/sKCSD.py
@@ -16,7 +16,7 @@ from scipy import special, interpolate, integrate
 from collections import Counter, OrderedDict
 import sys
 from .KCSD import KCSD1D
-from . import utility_functions as utils
+from . import sKCSD_utils as utils
 from . import basis_functions as basis
 
 class sKCSDcell(object):
